+ Added file_creation_date logic
+ This change set adds code to take the “creation time” of all the
files that are being uploaded and stores them in a dictionary .

+ This once again turned out to be more convoluted that initially
thought . the python.getctime() function does not give the actual
creation date of the file, it instead gives the time at which the meta
data for the files were created. So it always gives the last modified
time of the file as the creation time.

to circumvent this problem , I m using the bash command mdls ( metadata
list) and grepping FSCreationDate —> which gives the original file
creation date.

+ This dictionary is then stored persistently as creation.txt

=========================
Working on the following :
=========================
+ Create a set from the dict.value of the creation_date_dictionary and
mkdir for each of them

+ When the file is downloaded , it should be downloaded to the
appropriate buckets